Refactoring
Add id in mocks, remove empty fn in tests
