+++ src/components/artist-question-screen/artist-question-screen.e2e.test.js
@@ -13,14 +13,17 @@ const mock = {
       {
         artist: `one`,
         picture: `pic-one`,
+        id: 1234,
       },
       {
         artist: `two`,
         picture: `pic-two`,
+        id: 10098,
       },
       {
         artist: `three`,
         picture: `pic-three`,
+        id: 4567,
       },
     ],
   }
@@ -32,6 +35,7 @@ it(`Click on user answer should pass to the callback data-object from which this
   const userAnswer = {
     artist: `one`,
     picture: `pic-one`,
+    id: 1234,
   };
 
   const artistQuestion = shallow(
