+++ src/mocks/questions.js
@@ -7,15 +7,19 @@ export const questions = [
     answers: [{
       src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
       genre: `rock`,
+      id: Math.random(),
     }, {
       src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
       genre: `blues`,
+      id: Math.random(),
     }, {
       src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
       genre: `jazz`,
+      id: Math.random(),
     }, {
       src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
       genre: `rock`,
+      id: Math.random(),
     }],
   }, {
     type: `artist`,
@@ -26,12 +30,15 @@ export const questions = [
     answers: [{
       picture: `${AVATAR_URL}/${Math.random()}`,
       artist: `John Snow`,
+      id: Math.random(),
     }, {
       picture: `${AVATAR_URL}/${Math.random()}`,
       artist: `Jack Daniels`,
+      id: Math.random(),
     }, {
       picture: `${AVATAR_URL}/${Math.random()}`,
       artist: `Jim Beam`,
+      id: Math.random(),
     }],
   },
